Score evaluation, Tests and documentation
This commit refactors the documentation structure, and adds a bunch of extra information. I implemented the stubbs
for the score and scorer interfaces. A bit of head wrangling here as we're providing an function injection method on the scorer
this enables an interface method to be declared, and the ruleset can be implemented via anonmymous function injection.
Which is also actually rather cool ;-)